+++ src/setting.ts
@@ -2,10 +2,10 @@ import meta from "./meta";
 import { _updatePietem } from "./updating";
 
 function runNextFrameIfNeeded() {
-  if (meta.currentPhase === "waitingForFirstUpdate") {
+  if (meta.currentMetaPhase === "waitingForFirstUpdate") {
     meta.latestFrameId = requestAnimationFrame(_updatePietem);
 
-    meta.currentPhase = "waitingForMoreUpdates";
+    meta.currentMetaPhase = "waitingForMoreUpdates";
   }
 }
 
@@ -37,7 +37,7 @@ export function _setState(newState: any, callback?: any) {
     meta.mergeStates(
       typeof newState === "function" ? newState(meta.currentState) : newState,
       meta.currentState,
-      meta.currentPhase === "runningDeriveListeners"
+      meta.currentMetaPhase === "runningDeriveListeners"
         ? meta.recordedDeriveChanges
         : meta.recordedSubscribeChanges,
       meta.recordedSubscribeChanges